src/main/java/com/nray/nttback/dto/CustomerDTO.java
package com.nray.nttback.dto;

import lombok.Builder;

@Builder
public class CustomerDTO {

    private String documentType;
    private String documentNumber;
    private String firstName;
    private String middleName;
    private String lastName;
    private String secondLastName;
    private String phone;
    private String address;
    private String cityOfResidence;

    // Getters and Setters
    public String getDocumentType() {
        return documentType;
    }

    public String getDocumentNumber() {
        return documentNumber;
    }

    public String getFirstName() {
        return firstName;
    }

    public String getMiddleName() {
        return middleName;
    }

    public String getLastName() {
        return lastName;
    }

    public String getSecondLastName() {
        return secondLastName;
    }

    public String getPhone() {
        return phone;
    }

    public String getAddress() {
        return address;
    }

    public String getCityOfResidence() {
        return cityOfResidence;
    }
}
